Extracted redirects upload
