+++ core/server/api/v2/redirects.js
@@ -1,8 +1,6 @@
-const fs = require('fs-extra');
 const path = require('path');
-const moment = require('moment-timezone');
 const config = require('../../config');
-const validation = require('../../data/validation');
+
 const web = require('../../web');
 const redirects = require('../../../frontend/services/redirects');
 
@@ -28,37 +26,10 @@ module.exports = {
             cacheInvalidate: true
         },
         query(frame) {
-            const redirectsPath = path.join(config.getContentPath('data'), 'redirects.json');
-            const backupRedirectsPath = path.join(config.getContentPath('data'), `redirects-${moment().format('YYYY-MM-DD-HH-mm-ss')}.json`);
-
-            return fs.pathExists(redirectsPath)
-                .then((exists) => {
-                    if (!exists) {
-                        return null;
-                    }
-
-                    return fs.pathExists(backupRedirectsPath)
-                        .then((exists) => {
-                            if (!exists) {
-                                return null;
-                            }
-
-                            return fs.unlink(backupRedirectsPath);
-                        })
-                        .then(() => {
-                            return fs.move(redirectsPath, backupRedirectsPath);
-                        });
-                })
+            return redirects.handler.activate(frame.file.path)
                 .then(() => {
-                    return redirects.handler.readRedirectsFile(frame.file.path)
-                        .then((content) => {
-                            validation.validateRedirects(content);
-                            return fs.writeFile(redirectsPath, JSON.stringify(content), 'utf-8');
-                        })
-                        .then(() => {
-                            // CASE: trigger that redirects are getting re-registered
-                            web.shared.middlewares.customRedirects.reload();
-                        });
+                    // CASE: trigger that redirects are getting re-registered
+                    web.shared.middlewares.customRedirects.reload();
                 });
         }
     }
